WglGetProcAddress does not take UTF-16

+++ opengl32.go
@@ -211,7 +211,7 @@ func WglGetLayerPaletteEntries(hdc HDC, iLayerPlane, iStart, cEntries int, pcr *
 	return int(ret)
 }
 
-func WglGetProcAddress(lpszProc *uint16) uintptr {
+func WglGetProcAddress(lpszProc *byte) uintptr {
 	ret, _, _ := syscall.Syscall(wglGetProcAddress, 1,
 		uintptr(unsafe.Pointer(lpszProc)),
 		0,
